refactor(dashboard): change LoginAuthenticationFilter to an interface
